+++ libs/docs/core/vertical-navigation/vertical-navigation-header/vertical-navigation-header.component.ts
@@ -1,4 +1,7 @@
 import { Component } from '@angular/core';
+import { RouterLink } from '@angular/router';
+import { LinkComponent } from '@fundamental-ngx/core/link';
+import { MessageStripComponent } from '@fundamental-ngx/core/message-strip';
 import {
     DescriptionComponent,
     DocPageComponent,
@@ -11,6 +14,15 @@ import {
     selector: 'app-vertical-navigation-header',
     templateUrl: './vertical-navigation-header.component.html',
     standalone: true,
-    imports: [DocPageComponent, HeaderComponent, DescriptionComponent, ImportComponent, HeaderTabsComponent]
+    imports: [
+        DocPageComponent,
+        HeaderComponent,
+        DescriptionComponent,
+        ImportComponent,
+        HeaderTabsComponent,
+        MessageStripComponent,
+        LinkComponent,
+        RouterLink
+    ]
 })
 export class VerticalNavigationHeaderComponent {}